Refactor : block상수와 타입정의 blockTypes 파일로 분리
